Update README.md
Updated README.md to explain more on what the app does, as well as include thanks to Vitor Matias' medium article.

There is still some comments in Color.h that credit Vitor, as well as copies of the helper functions shown in the article. These should be removed in a future commit.